+++ python/polars_japanese/__init__.py
@@ -4,6 +4,8 @@ from . import common  # noqa: F401
 from .japanera_util import JapaneraExpr
 from .jpholiday_util import JpholidayExpr
 from .kanjize_util import KanjizeExpr
+from .normalize_util import NormalizeExpr
+from .prefecture import PrefectureExpr
 
 __version__ = version(__name__)
 
@@ -12,4 +14,5 @@ __all__ = [
     "KanjizeExpr",
     "JpholidayExpr",
     "NormalizeExpr",
+    "PrefectureExpr",
 ]